+++ src/window/glutin/src/headless.rs
@@ -15,7 +15,7 @@
 use std::os::raw::c_void;
 
 use device_gl::{Device, Factory, Resources as Res, create as gl_create, create_main_targets_raw};
-use glutin::{GlContext, HeadlessContext};
+use glutin::{NotCurrent, PossiblyCurrent, Context};
 
 use core::format::{Format, DepthFormat, RenderFormat};
 use core::handle::{DepthStencilView, RawDepthStencilView, RawRenderTargetView, RenderTargetView};
@@ -30,7 +30,7 @@ use core::texture::Dimensions;
 ///
 /// # Example
 ///
-/// ```rust
+/// ```
 /// extern crate gfx_core;
 /// extern crate gfx_window_glutin;
 /// extern crate glutin;
@@ -38,48 +38,49 @@ use core::texture::Dimensions;
 /// use gfx_core::format::{DepthStencil, Rgba8};
 /// use gfx_core::texture::AaMode;
 /// use gfx_window_glutin::init_headless;
-/// use glutin::HeadlessRendererBuilder;
+/// use glutin::{Context, ContextBuilder, EventsLoop};
 ///
 /// # fn main() {
 /// let dim = (256, 256, 8, AaMode::Multi(4));
 ///
-/// let context = HeadlessRendererBuilder::new(dim.0 as u32, dim.1 as u32)
-///     .build()
+/// let events_loop = EventsLoop::new();
+/// let context = ContextBuilder::new()
+///     .with_hardware_acceleration(Some(false))
+///     .build_headless(&events_loop, (256, 256).into())
 ///     .expect("Failed to build headless context");
 ///
-/// let (mut device, _, _, _) = init_headless::<Rgba8, DepthStencil>(&context, dim);
+/// let (context, device, factory, color, depth) = init_headless::<Rgba8, DepthStencil>(context, dim);
 /// # }
 /// ```
-pub fn init_headless<Cf, Df>(context: &HeadlessContext, dim: Dimensions)
-                             -> (Device, Factory,
+pub fn init_headless<Cf, Df>(context: Context<NotCurrent>, dim: Dimensions)
+                             -> (Context<PossiblyCurrent>, Device, Factory,
                                  RenderTargetView<Res, Cf>, DepthStencilView<Res, Df>)
     where
         Cf: RenderFormat,
         Df: DepthFormat,
 {
-    let (device, factory, color_view, ds_view) = init_headless_raw(context, dim,
+    let (ctx, device, factory, color_view, ds_view) = init_headless_raw(context, dim,
                                                                    Cf::get_format(),
                                                                    Df::get_format());
-    (device, factory, Typed::new(color_view), Typed::new(ds_view))
+    (ctx, device, factory, Typed::new(color_view), Typed::new(ds_view))
 }
 
 /// Raw version of [`init_headless`].
 ///
 /// [`init_headless`]: fn.init_headless.html
-pub fn init_headless_raw(context: &HeadlessContext, dim: Dimensions, color: Format, depth: Format)
-                         -> (Device, Factory,
+pub fn init_headless_raw(context: Context<NotCurrent>, dim: Dimensions, color: Format, depth: Format)
+                         -> (Context<PossiblyCurrent>, Device, Factory,
                              RawRenderTargetView<Res>, RawDepthStencilView<Res>)
 {
-    unsafe { context.make_current().unwrap() };
+    let context = unsafe { context.make_current().unwrap() };
 
-    let (device, factory) = gl_create(|s|
-        context.get_proc_address(s) as *const c_void);
+    let (device, factory) = gl_create(|s| context.get_proc_address(s) as *const c_void);
 
     // create the main color/depth targets
     let (color_view, ds_view) = create_main_targets_raw(dim, color.0, depth.0);
 
     // done
-    (device, factory, color_view, ds_view)
+    (context, device, factory, color_view, ds_view)
 }
 
 #[cfg(test)]
@@ -92,15 +93,17 @@ mod tests {
 
     #[test]
     fn test_headless() {
-        use glutin::{HeadlessRendererBuilder};
+        use glutin::{ContextBuilder, EventsLoop};
 
         let dim = (256, 256, 8, AaMode::Multi(4));
 
-        let context: HeadlessContext = HeadlessRendererBuilder::new(dim.0 as u32, dim.1 as u32)
-            .build()
+        let events_loop = EventsLoop::new();
+        let context = ContextBuilder::new()
+            .with_hardware_acceleration(Some(false))
+            .build_headless(&events_loop, (dim.0 as u32, dim.1 as u32).into())
             .expect("Failed to build headless context");
 
-        let (mut device, _, _, _) = init_headless::<Rgba8, DepthStencil>(&context, dim);
+        let (_, mut device, _, _, _) = init_headless::<Rgba8, DepthStencil>(context, dim);
 
         device.cleanup();
     }